remove strict temporarily
this is causing tests to fail, since it was added in node 9.9. Instead of increasing the node version required in package.json--which would necessitate a major bump in picomatch--it makes more sense to remove `strict` from assert. We're only testing strings here anyway. In the next major we'll add `strict` back.