Remove automatic windows detection, document `windows` option
To make this module usable in other environments besides Node.js, it
must not depend on node builtin modules. This PR does:

1. `process.platform` is replaced via the existing but undocumented
   `windows` option which is now documented.
2. `path.basename` is replaced by a simple regex replacement.

This is a breaking change for all users who are passing backslashes to
this module and they need to pass `process.platform === 'win32'` as the
`windows` option to attain previous behavior.

Fixes: https://github.com/micromatch/picomatch/issues/61
Fixes: https://github.com/micromatch/picomatch/issues/64
Fixes: https://github.com/micromatch/picomatch/issues/68